requirements.txt
flask
Pillow
ldap3